Fixed SceneWindow with no scene causing internal null pointer during draw (issue #150)
In case there's no scene connected we still want to invoke the connected compositor in order to clear the render target. The compositor however needs access to visibility groups which he gets from the scene instance. Therefor we use a dummy scene internally in case no scene is connected. This approach was easier then adding null checks inside of the stride rendering code.