Adds PolarCoordinates TextureFX.
Polar coordinates basically converts an incoming texture's mapping coordinates from cartesian space to circular (or polar) space or vice versa. The effect is the same as found in  the photoshop or gimp filter of the same name.
Textures seem to be a bit blurry compared to beta/dx11.
Maybe could be two separate nodes instead?