Use GitVersioning in our main repo only - a local dev build will have no version (or can be set manually with the normal msbuild properties if needed)
We currently reference the VL core projects directly when loading the projects inside our main solution, therefor the version can't be independent. Easiest solution is to treat the projects like if they would be hosted in our main repo by taking the version of the main repo. That way we avoid those super weired package reference conflicts we saw the past couple of days when exporting.