Removes the single sink restriction on nodes which feed values and resources into the shader graph
Each time a new shader graph gets generated a composite disposable is made available leaves can use to add (for example) a value subscription. That subscription is then tied to the sink itself (usually the effect instance which owns the parameter collection). This mechanism could also be used to tie other services to one graph.