Retrieve the parent manager of a component through an internal ConditionalWeakTable instead of S/R mechanism
This way we can be sure that all components get wrapped by a parent manager and will be attached to the entity itself.
In the previous approach this only worked for components wrapped by us - if one was using other components and didn't setup the required parent manager around it the component was not getting picked up by the Entity node.